third_party/WebKit/LayoutTests/external/wpt/html/rendering/non-replaced-elements/the-hr-element-0/setting-overflow-visible-expected.txt

This is a testharness.js-based test.
PASS control 
FAIL overflow: visible assert_equals: hr.offsetLeft expected 0 but got 50
Harness: the test ran to completion.
